SHRINKWRAP-332 moved ClassAsset and ClassLoaderAsset and their tests to API
