+++ api/src/main/java/org/jboss/shrinkwrap/api/asset/SecurityActions.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.shrinkwrap.impl.base.asset;
+package org.jboss.shrinkwrap.api.asset;
 
 import java.security.AccessController;
 import java.security.PrivilegedAction;
