+++ api/src/test/java/org/jboss/shrinkwrap/api/asset/ApiTestUtils.java
@@ -23,10 +23,8 @@ import java.io.InputStream;
 /**
  * ApiTestUtils
  *
- * Package private class that helps get byte array from {@link InputStream}.
- * Needed by the TestCases for common assets in api.
- *
  * @author <a href="mailto:ken@glxn.net">Ken Gullaksen</a>
+ * @author <a href="mailto:aslak@conduct.no">Aslak Knutsen</a>
  * @version $Revision: $
  */
 class ApiTestUtils
@@ -53,5 +51,55 @@ class ApiTestUtils
       out.close();
       in.close();
       return new String(out.toByteArray(), "UTF-8");
+   }
+   
+   /**
+    * Read the byte size of a {@link InputStream}.
+    * <br/>
+    * This method will close the stream when done. 
+    * 
+    * @param in Stream to get the size of.
+    * @return The byte size of the stream
+    * @throws Exception
+    */
+   static int findLengthOfStream(InputStream in) throws Exception 
+   {
+      int length = 0;
+      while(in.read() != -1) 
+      {
+         length++;
+      }
+      in.close();
+      return length;
+   }
+
+   /**
+    * Read the byte size of a {@link Class}.
+    * 
+    * @param clazz The class
+    * @return The byte size of the given {@link Class}
+    * @throws Exception
+    */
+   static int findLengthOfClass(Class<?> clazz) throws Exception 
+   {
+      String classResourceName = getResourceNameForClass(clazz);
+      InputStream in = SecurityActions.getThreadContextClassLoader().getResourceAsStream(classResourceName);
+      return findLengthOfStream(in);
+   }
+   
+   /**
+    * Get a resourceName for a {@link Class} so that it can be found in the {@link ClassLoader}.
+    * <br/>
+    * class.getName.relace( . -> / ) + ".class"
+    * 
+    * @param clazz The class to lookup
+    * @return
+    */
+   static String getResourceNameForClass(Class<?> clazz) 
+   {
+      String classResourceDelimiter = clazz.getName()
+                  .replaceAll("\\.", "/");
+      String classFullPath = classResourceDelimiter + ".class";
+      return classFullPath;
    }
 }
